// src/app/models/provider.ts
export class Provider {
    providersId: number;
    providerName: string;
    providerRtn: string;
    providerPhone1: string;
    providerPhone2: string;
    providerAddress: string;
    providerEmail: string;
    providerContact: string;
    contactPosition: string;
    city: Cities;
}

export class Cities {
    cityId: number;
    cityName: string;
}